feat: MailUtil 삭제 후 Service 통합
